requirements.txt
pandas
numpy
streamlit
langchain
langchain-community
weaviate-client
python-dotenv
pypdf
dataclasses